add project id for metrics controller and fix test.

+++ src/main/kotlin/com/thoughtworks/archguard/metrics/domain/MetricPersistService.kt
@@ -1,5 +1,5 @@
 package com.thoughtworks.archguard.metrics.domain
 
 interface MetricPersistService {
-    fun persistClassMetrics(projectId: Int)
+    fun persistClassMetrics(projectId: Long)
 }
